TRACK project details form in separate component to reuse
